refactor: moved validation functions into validations.ts
